+++ ribboncurls/tests/general.rs
@@ -200,3 +200,50 @@ fn data_with_nested_sections() {
 
     assert_eq!(template, "Tinted Theming!");
 }
+
+#[test]
+fn prefix_variable_with_dots_is_string() {
+    let template = "{{#a}}{{#b}}{{c.d.e.f}}{{/b}}{{/a}}";
+    let data = r#"
+        a:
+          c.d:
+            e:
+              f: Tinted Theming!
+        b: true
+    "#;
+    let template = ribboncurls::render(template, data, None).unwrap();
+
+    assert_eq!(template, "Tinted Theming!");
+}
+
+#[test]
+fn prefix_variable_with_dots_is_string_and_suffix_with_dots_is_string() {
+    let template = "{{#a}}{{#b}}{{c.d.e.f}}{{/b}}{{/a}}";
+    let data = r#"
+        a:
+          c.d:
+            e.f: Tinted Theming!
+        b: true
+    "#;
+    let template = ribboncurls::render(template, data, None).unwrap();
+
+    assert_eq!(template, "Tinted Theming!");
+}
+
+#[test]
+fn data_exists_in_parent_sections() {
+    let template = "{{#a}}{{#b}}{{data.exists}}{{moredata.doesntexist}}{{/b}}{{/a}}";
+    let data = r#"
+        a:
+          data:
+            exists: true
+        b:
+          moredata:
+            invalid: false
+        moredata:
+          true
+    "#;
+    let template = ribboncurls::render(template, data, None).unwrap();
+
+    assert_eq!(template, "true");
+}