+++ ribboncurls-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.2] - 2024-08-17
+
+### Changed
+
+- Update to latest `ribboncurls` dependency which includes bug fixes
+
 ## [0.3.1] - 2024-07-13
 
 ### Changed
@@ -31,6 +37,7 @@
 
 - Initial release
 
+[0.3.2]: https://github.com/tinted-theming/ribboncurls/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/tinted-theming/ribboncurls/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/tinted-theming/ribboncurls/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/tinted-theming/ribboncurls/compare/v0.2.0...v0.2.1